add provider address property, refactor subscribecontroller
