+++ src/main/resources/application-test.properties
@@ -37,6 +37,9 @@ rsk.blockchain.endpoint=http://localhost:7545
 rsk.blockchain.tokennetworkregistry=0xFEC354973ca22697BC5Cd1E7F372609574e2AfcA
 rsk.blockchain.multichaincontract=0xFEC354973ca22697BC5Cd1E7F372609574e2AfcA
 
+# Provider wallet address
+rif.notifier.provider.address=0x882bf23c4a7E73cA96AF14CACfA2CC006F6781A9
+
 # Rsk account
 rif.notifier.account.paassword=Z&VT=JRSAQAsMYgSXaCEMf!S7vXaDPfY+SW^aGC4TCATUTjExDb2+=GC7+fux88LdGtR!^m-v4dtX2_*2?9@E$7B7f!RCExn&4R9$GMHxK5vrk&qrp8aZ5Z4GzWPP5@vnffKpXU*2evJm!GtDCN-@BWZ-V*6EMeQ*3Rc=y-gbWFUGGm4+7y2j3fs@mxT$kSe48as-aTZ?BdtMUNgH4tJzeF2%hsy5$M@ELY86Z9dfLWWjq!Up-atxPr!Wt&3a2LZ
 rif.notifier.account.file=UTC--2019-04-19T15-07-00.568000000Z--034000b5f2862d114e4b3474f79fc64aad0cb742.json
